Restructure hypervisor file maps
The file maps on hypervisors have been structured, so instead of a
flat map, now there are two levels: cluster, and component. Below
component, there is a regular map with the filenames as keys, and
their SHA's as values.

This will make much easier the cleanup phase when deleting components,
when we want to remove files from the host.